requirements.txt
astroid==1.6.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
coveralls==1.3.0
docopt==0.6.2
flasgger==0.8.1
Flask==0.12.2
future==0.16.0
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nose==1.3.7
pycparser==2.18
pylint==1.8.2
python-dotenv==0.8.2
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
uuid==1.30
waitress==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
